backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
pydantic==2.10.3
pydantic-settings==2.7.0
alembic==1.13.2
